Double requests to remove a user now only require a second "OK".
It seems that when adding two users to a group and then removing the two
users, the second user will cause something like two button-press events.
The first one will properly remove the user, but the second one will fail
since the user is already removed from the group.

Now, if the user is already removed, a special 506 status code is returned
which the Javascript interprets as "User is already out of the group" and
hides the user element from display.